Add --has-inner-blocks option to scaffold block command.
By passing the `--has-inner-blocks` flag, the save and edit functions will be set up to use `innerBlocks`. This also forces `isDynamic` to be `false`, by definition. To enable this change,
it also means that the default option for the `isDynamic` option is also `false` unless the flag is explicitly passed to the command.